backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
pydantic==2.5.0
python-multipart==0.0.6
redis==5.0.1
requests==2.31.0
celery==5.3.4
python-dotenv==1.0.0
bcrypt==4.0.1